feat: move directoy services and model to mangar directory & added new methods(create, update and delete)
