--- src/app/manager/invoices/model/customer.entity.spec.ts
@@ -1,7 +1,0 @@
-import { CustomerEntity } from './customer.entity';
-
-describe('CustomerEntity', () => {
-  it('should create an instance', () => {
-    expect(new CustomerEntity()).toBeTruthy();
-  });
-});
