--- src/app/manager/invoices/model/invoice.entity.spec.ts
@@ -1,7 +1,0 @@
-import { InvoiceEntity } from './invoice.entity';
-
-describe('InvoiceEntity', () => {
-  it('should create an instance', () => {
-    expect(new InvoiceEntity()).toBeTruthy();
-  });
-});
