--- src/app/manager/invoices/model/vehicle.entity.spec.ts
@@ -1,7 +1,0 @@
-import { VehicleEntity } from './vehicle.entity';
-
-describe('VehicleEntity', () => {
-  it('should create an instance', () => {
-    expect(new VehicleEntity()).toBeTruthy();
-  });
-});
