--- src/app/manager/invoices/model/workshop.entity.spec.ts
@@ -1,7 +1,0 @@
-import { WorkshopEntity } from './workshop.entity';
-
-describe('WorkshopEntity', () => {
-  it('should create an instance', () => {
-    expect(new WorkshopEntity()).toBeTruthy();
-  });
-});
