--- src/app/manager/invoices/services/vehicle.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { VehicleService } from './vehicle.service';
-
-describe('VehicleService', () => {
-  let service: VehicleService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(VehicleService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
