+++ src/app/manager/model/model/customer.entity.spec.ts
@@ -1,0 +1,7 @@
+import { CustomerEntity } from './customer.entity';
+
+describe('CustomerEntity', () => {
+  it('should create an instance', () => {
+    expect(new CustomerEntity()).toBeTruthy();
+  });
+});
