+++ src/app/manager/model/model/invoice.entity.spec.ts
@@ -1,0 +1,7 @@
+import { InvoiceEntity } from './invoice.entity';
+
+describe('InvoiceEntity', () => {
+  it('should create an instance', () => {
+    expect(new InvoiceEntity()).toBeTruthy();
+  });
+});
