+++ src/app/manager/model/model/vehicle.entity.spec.ts
@@ -1,0 +1,7 @@
+import { VehicleEntity } from './vehicle.entity';
+
+describe('VehicleEntity', () => {
+  it('should create an instance', () => {
+    expect(new VehicleEntity()).toBeTruthy();
+  });
+});
