+++ src/app/manager/model/model/workshop.entity.spec.ts
@@ -1,0 +1,7 @@
+import { WorkshopEntity } from './workshop.entity';
+
+describe('WorkshopEntity', () => {
+  it('should create an instance', () => {
+    expect(new WorkshopEntity()).toBeTruthy();
+  });
+});
